Extract FilterSlider to new module
